% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_simulate_ml_Hway_suffstats}
\alias{rcpp_simulate_ml_Hway_suffstats}
\title{rcpp_simulate_ml_Hway_suffstats}
\usage{
rcpp_simulate_ml_Hway_suffstats(
  rsamp_num,
  rburnin,
  rinterval,
  rmodel_dim,
  model_terms,
  rnum_nodes,
  rnum_layers,
  highest_order,
  rtheta,
  rand_seed,
  g
)
}
\description{
Simulate a multilayer network model and print out sufficient statistics
}
\examples{
rcpp_simulate_ml_Hway_suffstats(1,100,1,2,c("withinlayer","crosslayer"),3,2,2, c(0.4,0.7),123,1)

}
